refactor: change memory repository directory
