+++ internal/repository/memory.go
@@ -1,4 +1,4 @@
-package memory
+package repository
 
 import (
 	"context"
